move tests::repository::config::worktree to `tests::repository::open::worktree`
As these settings are relevant during opening and aren't queried later
as part of implementations in `repository::config` or `config`.
