+++ git-repository/tests/repository/config/mod.rs
@@ -3,4 +3,3 @@ mod identity;
 mod remote;
 #[cfg(any(feature = "blocking-network-client", feature = "async-network-client"))]
 mod transport_options;
-mod worktree;
